backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
DAWG-Python==0.7.2
defusedxml==0.8.0rc2
Django==5.0.4
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
ffmpeg==1.4
gensim==4.3.2
idna==3.7
joblib==1.4.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
psycopg2==2.9.9
pycparser==2.22
pydub==0.25.1
PyJWT==2.8.0
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
python-dotenv==1.0.1
python3-openid==3.2.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
scipy==1.12.0
smart-open==7.0.4
social-auth-app-django==5.4.0
social-auth-core==4.5.3
SpeechRecognition==3.10.3
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
